Fixed apprequests dialog support on Android
The previous implementation of DialogActivity assumed a feed dialog and processed the response accordingly. The iOS version does not have this limitation.

This update tracks the type of dialog invoked (either "feed", or "apprequests" - which are the only types supported by FB at the moment).

This fix is backwards compatible.